refactor(web): move humanizeRequirement to apps/web utils
The Stripe Connect requirement humaniser is a UI concern (renders
field paths for the monetisation page), not a subscription-domain
primitive. Move it out of @codex/subscription into apps/web utils
under the more specific name connect-requirement-humanization so
the package boundary reflects responsibility.
