+++ apps/web/src/lib/utils/connect-requirement-humanization.ts
